% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gghist-package.R
\docType{package}
\name{gghist-package}
\alias{gghist-package}
\alias{_PACKAGE}
\title{gghist: Plot the Histogram of a Numeric Vector}
\description{
Wrapper around geom_histogram() of 'ggplot2' to plot the histogram of a numeric vector. This is especially useful, since qplot() was deprecated in 'ggplot2' 3.4.0.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/frederikziebell/gghist}
}

}
\author{
\strong{Maintainer}: Frederik Ziebell \email{frederik.ziebell@gmail.com} (\href{https://orcid.org/0000-0003-3673-1721}{ORCID})

}
\keyword{internal}
