% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gghist.R
\name{gghist}
\alias{gghist}
\title{Plot the histogram of a numeric vector}
\usage{
gghist(x, ...)
}
\arguments{
\item{x}{A numeric vector}

\item{...}{Other arguments passed on to geom_histogram()}
}
\value{
A ggplot object
}
\description{
This is a wrapper around \code{ggplot2::geom_histogram()} to display the histogram of a numeric vector
}
\examples{
x <- rnorm(100)
gghist(x, bins = 20)
}
